add a tardy transaction eliminator
It works well enough for clients that've been indefinitely pre-empted
regardless of whether they're inside a transaction or not. Clients in a
transaction will observe an -ETIMEDOUT status from xn_commit().

This adds another blind spot in xn_commit() where pre-emption may delay epoch
collection indefinitely: it's between the TXNID_COMMIT setting fetch-and-or
and the clean_client() call in xn_commit(). Since this is also the epoch
consumer segment, the prevention effect is unavoidable unless the thread can
be re-killed at pre-empt.

(the other blindspot is the epoch_lock flag, where pre-emption in the current
holder delays collection similarly.)

This changeset also generates the commit ID only after it's known that the
transaction wasn't killed, and won't be restarted due to snapshot violation.
Doing so conserves transaction IDs (big deal, right? there's 16M of 'em.) in
transactions that died young.